refactor: chain
